requirements.txt
PyMySQL~=1.0.2
pandas~=1.1.5
SQLAlchemy==1.4.2
python-dotenv~=0.15.0
PyPika~=0.48.0
numpy~=1.19.5
tensorflow-cpu~=2.5.0
wheel
gym